sccp: Fix test rx SLS expectancies
The SLS is the same for all messages in conn being sent in one direction, but
doesn't need to be the same value on both directions.
Since the SLS value on the other direction is not selected by the test
itself, we cannot expect a given specifi value.
Update the test expectancies.

This started to fail since recently libosmo-sigtran started properly
setting SLS values, eg libosmo-sigtran.git
7781eb275da41a9b6b1ea5d8b0e802e87a8e9d53 and
0061e8d0bcba3b0ed5ea255588619627d0975380.

Change-Id: Ic288b58629361de9cfadb568a27b51b7e02c9e99