testenv: clone_project: fix getting latest version
Extend the logic for getting the last version, so it doesn't only work
with libosmo-sigtran (where the last version happens to be the last one
returned by "git ls-remote --tags") but also for libosmocore where this
isn't the case. Filter the versions by the relevant ones and then sort
them to get the highest one.

Change-Id: Ic15e385b9c8bab5c0dc70276049d6ad5ae927a6a