stp: Initial split into several testenv config
As a first step, split the generic config into 3 configs, one per
module.

The idea is to have a different config for a given set of related tests
running on a specific network topology/setup.

This helps in simplifying the test conditions, which allows easily
adding new topologies to test, validate specific scenarios or debug
existing ones.

For instance, to test Combined linkset loadsharing, we'll need to set up
some sort of scenario where we have at least 2 AS with each at least 1
ASP.

Change-Id: I4ec8e1bbc8e71d8bd421dfbd5356bf1937ce2bb4