epdg: TC_concurrent_ues_MTC: rx CMD_ATTACHED without from vc_conn_list[i]
This is needed to make the test pass in QEMU. I think this makes it
accept the COORD_CMD_ATTACHED in any order instead of the order the
COORD_CMD_START were sent beforehand.

Increasing the sleep between the COORD_CMD_START also makes the test
pass in QEMU, but that increases the overall test time significantly in
TC_concurrent_ues100.

Change-Id: I4c76d12273a14df562ea906e6c8ddb348d1b0eda