testenv: run "make deps" every time
The previous logic of not always running "make deps" was useful when
"make deps" would take a long time to complete even if there were no
changes. But this has been fixed in c7e40d6a ("Makefile: avoid fetching
deps if not needed"). Run it every time now in order to fix that deps
are not updated after deps/Makefile gets changed.

Change-Id: I02668530ce36c51a1c641fb816464c9c8e6ded9e