testenv: enable_binary_repo: configure apt pin
Add an apt-pin to ensure packages from the Osmocom repositories get
preferred over Debian's packages. I just ran into this with osmo-mgw,
where we have 1.4.0 in osmocom:latest and Debian 13 has 1.14.0+dfsg1-2,
which counts as higher version. This results in the wrong package being
selected, and also in an error later on when trying to install related
debug packages from our repository:

  The following packages have unmet dependencies:
   osmo-mgw-dbgsym : Depends: osmo-mgw (= 1.14.0) but 1.14.0+dfsg1-2 is to be installed

Change-Id: I3a45de277bbd5299c8b840680b56b52a8731caa4