ranap: Support Tx GlobalRNC-ID during RANAP RESET (ACK)
This commit doesn't affect the behavior of any existing testsuite yet,
it only updates the templates and adds a RanOps param to be able to pass
a GlobalRNC-Id to RAN_Emulation.

A follow-up commit will change SGSN_Tests to append GlobalRNC-Id to
RANAP Reset when sending towards SGSN, since it's mandatory for an RNC
to set it according to 3GPP TS 25.413 8.26.2.2.

Change-Id: I0628e412bdff6ed8eefe9051ec351fbd1d0b634f