SGsAP_Emulation: Introduce proper support for server-mode
Prior to this commit, only initial (not really useful) SCTP server-mode
support existed in SGsAP_Emulation.
This is basically because MSC_Tests uses the SCTP client-mode, and
MME_Tests_SGsAP were not yet in a fully working state.

In server-mode, we have a conn_id for the listening socket and a conn_id
for the accepted socket; track them properly.
Moreover, in server node we definetly need some sort of event to wait
for the client to connect; introduce it.
Based on existing work in Iuh_Emulation.

Change-Id: Iccf4ac96c56e947529f0ffc06428e2325a115d50