stp: define PC for IPA ASP peers
This allows explicitly matching what's configured in osmo-stp's
"routing-key" & "point-code override dpc" VTY configs for each AS.
This in turn makes the osmo-stp cfg file easy to understand by looking
at the peer's (TTCN3) configuration.

There's no need to configure "point-code override opc" in the osmo-stp
since by default the PC declared in the routing-key is used to apply an
OPC on incoming messages over an IPA ASP.

Once we have the PC available in the TTCN3 config file, we can use it
inside IPA<->M3UA test to put the value in generated M3UA messages
(there's no visible changes in IPA ones because ofc the IPA signalling
doesn't include the PC in the messages).

Change-Id: I9cca2f4dd38e03ab1d259ffafd904b8890778883