RAN_{Emulation,Adapter}: Split transport type from transport client/server mode
The RAN_Transport is used to identify the stack of protocols in use up
to SCCP. The client/server mode is some extra independent information.
For instance, RAN_Emulation doesn't really use it while RAN_Adapter
does, since the later is the one in charge of setting up the lower layer
socket connection.

Change-Id: Ib2e7d029bd7eb59cc6f3b3a28bc417f00d3fed9c