testenv: coredump: support core_pattern=core
Support getting core files from a typical core_pattern=core where the
coredump just gets stored in the current working dir, instead of always
retrieving it from coredumpctl. This is what we will use with jenkins in
the future, as it makes getting core files in other jobs easier. Remove
support for the custom testenv-coredump-helper code that isn't needed
anymore.

Change-Id: Ia765b01432e4cb4cd36c45de874b966e3ebf55bc