hlr: pyhss: create/delete subscribers
Run the PyHSS API service, and fill it with a default APN on startup.
Having one APN entry in the database is required for creating
subscribers.

Talk to the API service for creating and deleting subscribers. Do this
with a new script pyhss_api_helper.py. Deleting subscribers requires
reading JSON returned from the server (to get the AUC and subscriber ID
from the IMSI). I have first attempted to do this via HTTP_Adapter
instead of using a helper script, but this was a lot more complex and
would have required to have the JSON structure in the TTCN3 files. The
eim testsuite also runs an external script for REST requests.

With this change and additional fixes in PyHSS, more tests pass:
* HLR_Tests.TC_gsup_sai
* HLR_Tests.TC_gsup_sai_num_auth_vectors
* HLR_Tests.TC_gsup_ul
* HLR_Tests.TC_gsup_purge_cs
* HLR_Tests.TC_gsup_purge_ps

Related: OS#6862
Change-Id: Ic924dabbc813459f73d6646ee17b79cb11d39a76