smdpp_Tests: simplify generation of smdpp_ConnHdlrPars (NIST vs. BRP)
When we call f_init_pars() it will return a record with default parameters.
However, those parameters are only a suitable default when the NIST curve
is used. For the brainpool curve we will have to modify returned result.

Let's simlify this by adding a parameter to f_init_pars() that allows us to
switch between the defaults for NIST and BRP as needed.

Realated: SYS#7339
Change-Id: Ied4ffa6ff4b3efbbcfc9204ce8f21ea7ca4becf5