ES2Plus_Tests: rename mp_es2plus_server_host to _fqdn
Let's rename the module parameter mp_es2plus_server_host to
mp_es2plus_server_fqdn. The reason for this is that the suffix "host"
suggests that one may use an arbitrary hostname or IP address. In
this case this is in fact not true. The FQDN must match the hostname
of the SSL server certificate. Apart from that GSMA SGP.22 always
uses the word "FQDN" in their specs.

Change-Id: Idedfcce4d013c32f43307e159c45854a1e6a049b
Related: SYS#7339