smdpp_Tests.cfg: comment out parameters in config file.
The code in smdpp_Tests.ttcn already provides safe defaults, so we
do not have to repeat those defaults again in smdpp_Tests.cfg.

So let's comment out the parameters. This way we kenn keep them for
illustrative purposes, so that users can use the config file as a
helpful template for the most basic settings.

Related: SYS#7339
Change-Id: I981b683da0312799c2e5b24ac57cd838b071f955