smdpp_Tests: optimize calling of f_init_esXplus
When we call the f_init_esXplus functions we always check the return
code to see if the initialization went ok. However in both functions
we already set the verdict and in the ES2+ variant of the functions
we already stop the MTC. Let's optimize the function calls so that
we only call the function without checking the return code. Let's
do all the error handling and the stopping of the MTC in inside
f_init_esXplus.

Related: SYS#7339
Change-Id: I64fc5a7eddd6c5aee8ce85dbfe56c076f3372f3d