StatsD_Checker: Abort on expectancy failure with clear status message
Before this commit, a statsd expect failure would let the test continue
the test for a while and then finally log a rellay generic timeout
message, which makes really difficult to understand what's going on and
why is the test failing.

Let's terminate the test early if a test expectation fails, so it's a
lot easier to debug.
Add an extra param to be able to use the API for test logic to check the
state without failing/aborting. This is not really used anywhere so far
but the API was built this way and I indeed think it may be handy in the
future.

Change-Id: Ibc3d079c2a45aa22f9e2e42d75109cae66530f86